+++ src/DatePicker/__tests__/DatePicker-test.js
@@ -11,9 +11,12 @@ describe('DatePicker', () => {
   })
 
   it('should onSelect works', () => {
+    document.body.innerHTML = ''
     const handleSelect = jest.fn()
     const instance = TestUtils.renderIntoDocument(<DatePicker date="2016-01-01" onSelect={handleSelect} />)
-    TestUtils.Simulate.click(findDOMNode(instance).querySelector('tbody button'))
+    const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+    TestUtils.Simulate.click(document.querySelector('tbody button'))
     expect(handleSelect).toBeCalled()
   })
 })