+++ src/DatePicker/__tests__/DateRange-test.js
@@ -5,9 +5,12 @@ import DateRange from '../DateRange'
 
 describe('DateRange', () => {
   it('should onSelect works', () => {
+    document.body.innerHTML = ''
     const handleSelect = jest.fn()
     const instance = TestUtils.renderIntoDocument(<DateRange start="2016-01-01" onSelect={handleSelect} />)
-    TestUtils.Simulate.click(findDOMNode(instance).querySelector('tbody button'))
+    const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+    TestUtils.Simulate.click(document.querySelector('tbody button'))
     expect(handleSelect).toBeCalled()
   })
 })