+++ src/MultipleSelect/__tests__/index-test.js
@@ -5,6 +5,10 @@ import { MultipleSelect, Option } from '../index'
 
 describe('MultipleSelect', () => {
 
+  beforeEach(() => {
+    document.body.innerHTML = ''
+  })
+
   it('should values works', () => {
     const instance = TestUtils.renderIntoDocument(
       <MultipleSelect values={['1', '2']} onChange={jest.fn()}>
@@ -13,8 +17,11 @@ describe('MultipleSelect', () => {
         <Option value="2">小米</Option>
       </MultipleSelect>
     )
-    const checkboxes = findDOMNode(instance).querySelectorAll('.bfd-checkbox input:checked')
-    const labels = findDOMNode(instance).querySelectorAll('.bfd-tag-list__item')
+    const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+
+    const checkboxes = document.querySelectorAll('.bfd-checkbox input:checked')
+    const labels = container.querySelectorAll('.bfd-tag-list__item')
     expect(checkboxes.length).toBe(2)
     expect(labels.length).toBe(2)
   })
@@ -28,7 +35,10 @@ describe('MultipleSelect', () => {
         <Option value="2">小米</Option>
       </MultipleSelect>
     )
-    const checkboxes = findDOMNode(instance).querySelectorAll('.bfd-checkbox input:checked')
+    const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+
+    const checkboxes = document.querySelectorAll('.bfd-checkbox input:checked')
     expect(checkboxes.length).toBe(2)
   })
 
@@ -42,24 +52,26 @@ describe('MultipleSelect', () => {
       </MultipleSelect>
     )
     const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+
     TestUtils.Simulate.change(container.querySelector('input'), {
       target: {
         value: '1'
       }
     })
-    expect(container.querySelectorAll('.bfd-checkbox').length).toBe(2)
+    expect(document.querySelectorAll('.bfd-checkbox').length).toBe(2)
     TestUtils.Simulate.change(container.querySelector('input'), {
       target: {
         value: '小'
       }
     })
-    expect(container.querySelectorAll('.bfd-checkbox').length).toBe(2)
+    expect(document.querySelectorAll('.bfd-checkbox').length).toBe(2)
     TestUtils.Simulate.change(container.querySelector('input'), {
       target: {
         value: 'x'
       }
     })
-    expect(container.querySelectorAll('.bfd-checkbox').length).toBe(0)
+    expect(document.querySelectorAll('.bfd-checkbox').length).toBe(0)
   })
 
   it('should onChange works', () => {
@@ -71,7 +83,10 @@ describe('MultipleSelect', () => {
         <Option value="2">小米</Option>
       </MultipleSelect>
     )
-    const checkboxes = findDOMNode(instance).querySelectorAll('.bfd-checkbox input')
+    const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+
+    const checkboxes = document.querySelectorAll('.bfd-checkbox input')
     TestUtils.Simulate.change(checkboxes[1], {
       target: {
         checked: true
@@ -85,7 +100,10 @@ describe('MultipleSelect', () => {
     const instance = TestUtils.renderIntoDocument(
       <MultipleSelect data={['a', 'b']} render={item => <Option>{item}</Option>} />
     )
-    const checkboxes = findDOMNode(instance).querySelectorAll('.bfd-checkbox')
+    const container = findDOMNode(instance)
+    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+
+    const checkboxes = document.querySelectorAll('.bfd-checkbox')
     expect(checkboxes.length).toBe(3)
   })
 
@@ -95,21 +113,25 @@ describe('MultipleSelect', () => {
       <MultipleSelect tagable />
     )
     const container = findDOMNode(instance)
-    TestUtils.Simulate.change(container.querySelector('input'), {
+    const searchInput = container.querySelector('input')
+
+    TestUtils.Simulate.change(searchInput, {
       target: {
         value: 'a'
       }
     })
-    TestUtils.Simulate.keyDown(container.querySelector('input'), {
+    TestUtils.Simulate.keyDown(searchInput, {
       key: 'Enter'
     })
-    const labels = findDOMNode(instance).querySelectorAll('.bfd-tag-list__item')
+    const labels = container.querySelectorAll('.bfd-tag-list__item')
     expect(labels.length).toBe(1)
   })
 
   it('should empty children works', () => {
-    TestUtils.renderIntoDocument(
-      <MultipleSelect>{null}</MultipleSelect>
-    )
+    expect(() => {
+      TestUtils.renderIntoDocument(
+        <MultipleSelect>{null}</MultipleSelect>
+      )
+    }).not.toThrow()
   })
 })